{"schema":"decodex/commit/1","summary":"Split docs source capture helpers after strict validation.","authority":"manual"}
